requirements.txt
flask
flask-mysql
flask-wtf
flask-bcrypt
flask-login
flask-migrate
pymysql
email-validator
python-dotenv
cryptography